112 Mock Mongo and Elastic Search Rest client which would otherwise leak threads in tests
