+++ test/controllers/ReportGeneratorWizardController_AutoSave_Test.java
@@ -1,12 +1,14 @@
 package controllers;
 
 import com.google.common.collect.ImmutableMap;
+import com.mongodb.rx.client.MongoClient;
 import data.ShortFormatPreSentenceReportData;
 import helpers.Encryption;
 import interfaces.AnalyticsStore;
 import interfaces.DocumentStore;
 import interfaces.PdfGenerator;
 import lombok.val;
+import org.elasticsearch.client.RestHighLevelClient;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -214,7 +216,9 @@ public class ReportGeneratorWizardController_AutoSave_Test extends WithApplicati
             overrides(
                 bind(PdfGenerator.class).toInstance(pdfGenerator),
                 bind(DocumentStore.class).toInstance(alfrescoDocumentStore),
-                bind(AnalyticsStore.class).toInstance(analyticsStore)
+                bind(AnalyticsStore.class).toInstance(analyticsStore),
+                bind(RestHighLevelClient.class).toInstance(mock(RestHighLevelClient.class)),
+                bind(MongoClient.class).toInstance(mock(MongoClient.class))
             )
             .build();
     }
