+++ test/views/OffenderAssessmentWebTest.java
@@ -2,19 +2,12 @@ package views;
 
 import com.google.common.collect.ImmutableMap;
 import helpers.JwtHelperTest;
-import interfaces.AnalyticsStore;
-import interfaces.DocumentStore;
-import interfaces.OffenderApi;
-import interfaces.PdfGenerator;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
-import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
-import play.Application;
-import play.inject.guice.GuiceApplicationBuilder;
 import play.libs.Json;
 import views.pages.CheckYourReportPage;
 import views.pages.OffenderAssessmentPage;
@@ -29,19 +22,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.data.MapEntry.entry;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Matchers.any;
-import static org.mockito.Mockito.atLeastOnce;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static play.inject.Bindings.bind;
+import static org.mockito.Mockito.*;
 import static utils.CourtAppearanceHelpers.someCourtAppearances;
 import static utils.OffenderHelper.anOffenderWithNoContactDetails;
 import static views.helpers.AlfrescoDataHelper.legacyReportWith;
 
 @RunWith(MockitoJUnitRunner.class)
 public class OffenderAssessmentWebTest extends WithIE8Browser {
-    @Mock
-    private DocumentStore alfrescoDocumentStore;
     @Captor
     private ArgumentCaptor<String> metaDataCaptor;
 
@@ -64,10 +51,14 @@ public class OffenderAssessmentWebTest extends WithIE8Browser {
         offenderAssessmentPage = new OffenderAssessmentPage(browser);
         startPage = new StartPage(browser);
         checkYourReportPage = new CheckYourReportPage(browser);
-        when(alfrescoDocumentStore.updateExistingPdf(any(), any(), any(), any(), any()))
+        when(documentStore.updateExistingPdf(any(), any(), any(), any(), any()))
                 .thenReturn(CompletableFuture.completedFuture(ImmutableMap.of("ID", "123")));
-        when(alfrescoDocumentStore.uploadNewPdf(any(), any(), any(), any(), any(), any()))
+        when(documentStore.uploadNewPdf(any(), any(), any(), any(), any(), any()))
                 .thenReturn(CompletableFuture.completedFuture(ImmutableMap.of("ID", "123")));
+        given(offenderApi.logon(any())).willReturn(CompletableFuture.completedFuture(JwtHelperTest.generateToken()));
+        given(offenderApi.getOffenderByCrn(any(), any())).willReturn(CompletableFuture.completedFuture(anOffenderWithNoContactDetails()));
+        given(offenderApi.getCourtAppearancesByCrn(any(), any())).willReturn(CompletableFuture.completedFuture(someCourtAppearances()));
+        given(pdfGenerator.generate(any(), any())).willReturn(CompletableFuture.supplyAsync(() -> new Byte[0]));
     }
 
     @Test
@@ -130,7 +121,7 @@ public class OffenderAssessmentWebTest extends WithIE8Browser {
 
     @Test
     public void legacyReportWithDrugsIssueBecomesSubstanceMisuse() {
-        when(alfrescoDocumentStore.retrieveOriginalData(any(), any())).
+        when(documentStore.retrieveOriginalData(any(), any())).
                 thenReturn(legacyReportWith(
                         ImmutableMap.of("issueDrugs", "true", "pageNumber", "7")));
 
@@ -142,7 +133,7 @@ public class OffenderAssessmentWebTest extends WithIE8Browser {
 
     @Test
     public void legacyReportWithAlcoholIssueDetailsBecomesOtherIssueWithDetails() {
-        when(alfrescoDocumentStore.retrieveOriginalData(any(), any())).
+        when(documentStore.retrieveOriginalData(any(), any())).
                 thenReturn(legacyReportWith(
                         ImmutableMap.of("issueAlcohol", "true", "pageNumber", "7")));
 
@@ -154,7 +145,7 @@ public class OffenderAssessmentWebTest extends WithIE8Browser {
 
     @Test
     public void legacyReportWithAlcoholIssueBecomesSubstanceMisuse() {
-        when(alfrescoDocumentStore.retrieveOriginalData(any(), any())).
+        when(documentStore.retrieveOriginalData(any(), any())).
                 thenReturn(legacyReportWith(
                         ImmutableMap.of("offenderAssessment", "some offender assessment", "pageNumber", "7")));
 
@@ -167,7 +158,7 @@ public class OffenderAssessmentWebTest extends WithIE8Browser {
 
     @Test
     public void existingReportWithNoIssuesHasNothingTickedAndNoDetails() {
-        when(alfrescoDocumentStore.retrieveOriginalData(any(), any())).
+        when(documentStore.retrieveOriginalData(any(), any())).
                 thenReturn(legacyReportWith(
                         ImmutableMap.<String, Object>builder().
                                 put("pageNumber", "7").
@@ -197,7 +188,7 @@ public class OffenderAssessmentWebTest extends WithIE8Browser {
 
     @Test
     public void existingReportWithAllIssuesHasEverythingTickedWithDetails() {
-        when(alfrescoDocumentStore.retrieveOriginalData(any(), any())).
+        when(documentStore.retrieveOriginalData(any(), any())).
                 thenReturn(legacyReportWith(
                         ImmutableMap.<String, Object>builder().
                                 put("pageNumber", "7").
@@ -226,7 +217,7 @@ public class OffenderAssessmentWebTest extends WithIE8Browser {
     }
 
     private Map<String, String> storedData() {
-        verify(alfrescoDocumentStore, atLeastOnce()).updateExistingPdf(any(), any(), any(), metaDataCaptor.capture(), any());
+        verify(documentStore, atLeastOnce()).updateExistingPdf(any(), any(), any(), metaDataCaptor.capture(), any());
         return jsonToMap(Json.parse(metaDataCaptor.getValue()).get("values"));
     }
 
@@ -247,26 +238,6 @@ public class OffenderAssessmentWebTest extends WithIE8Browser {
     private void givenAllIssuesAreTicked() {
         stream(issueOptions).forEach(issue -> offenderAssessmentPage.tick(issue));
         stream(issueOptions).forEach(issue -> offenderAssessmentPage.fillDetailsWith(issue, issue + " details"));
-    }
-
-    @Override
-    protected Application provideApplication() {
-        PdfGenerator pdfGenerator = mock(PdfGenerator.class);
-        given(pdfGenerator.generate(any(), any())).willReturn(CompletableFuture.supplyAsync(() -> new Byte[0]));
-
-        OffenderApi offenderApi = mock(OffenderApi.class);
-        given(offenderApi.logon(any())).willReturn(CompletableFuture.completedFuture(JwtHelperTest.generateToken()));
-        given(offenderApi.getOffenderByCrn(any(), any())).willReturn(CompletableFuture.completedFuture(anOffenderWithNoContactDetails()));
-        given(offenderApi.getCourtAppearancesByCrn(any(), any())).willReturn(CompletableFuture.completedFuture(someCourtAppearances()));
-
-        return new GuiceApplicationBuilder().
-            overrides(
-                bind(PdfGenerator.class).toInstance(pdfGenerator),
-                bind(DocumentStore.class).toInstance(alfrescoDocumentStore),
-                bind(AnalyticsStore.class).toInstance(mock(AnalyticsStore.class)),
-                bind(OffenderApi.class).toInstance(offenderApi))
-            .configure("params.user.token.valid.duration", "100000d")
-            .build();
     }
 
 }
