+++ test/views/SaveAsDraftWebTest.java
@@ -2,17 +2,10 @@ package views;
 
 import com.google.common.collect.ImmutableMap;
 import helpers.JwtHelperTest;
-import interfaces.AnalyticsStore;
-import interfaces.DocumentStore;
-import interfaces.OffenderApi;
-import interfaces.PdfGenerator;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
-import play.Application;
-import play.inject.guice.GuiceApplicationBuilder;
 import views.pages.CheckYourReportPage;
 import views.pages.DraftSavedConfirmationPage;
 import views.pages.OffenderAssessmentPage;
@@ -23,19 +16,12 @@ import java.util.concurrent.CompletableFuture;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Matchers.any;
-import static org.mockito.Mockito.atLeastOnce;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static play.inject.Bindings.bind;
+import static org.mockito.Mockito.*;
 import static utils.CourtAppearanceHelpers.someCourtAppearances;
 import static utils.OffenderHelper.anOffenderWithNoContactDetails;
 
 @RunWith(MockitoJUnitRunner.class)
 public class SaveAsDraftWebTest extends WithIE8Browser {
-    @Mock
-    private DocumentStore alfrescoDocumentStore;
-
     private OffenderAssessmentPage offenderAssessmentPage;
     private OffenderDetailsPage offenderDetailsPage;
     private DraftSavedConfirmationPage draftSavedConfirmationPage;
@@ -47,10 +33,17 @@ public class SaveAsDraftWebTest extends WithIE8Browser {
         offenderDetailsPage = new OffenderDetailsPage(browser);
         draftSavedConfirmationPage = new DraftSavedConfirmationPage(browser);
         checkYourReportPage = new CheckYourReportPage(browser);
-        when(alfrescoDocumentStore.updateExistingPdf(any(), any(), any(), any(), any()))
+        when(documentStore.updateExistingPdf(any(), any(), any(), any(), any()))
                 .thenReturn(CompletableFuture.completedFuture(ImmutableMap.of("ID", "123")));
-        when(alfrescoDocumentStore.uploadNewPdf(any(), any(), any(), any(), any(), any()))
+        when(documentStore.uploadNewPdf(any(), any(), any(), any(), any(), any()))
                 .thenReturn(CompletableFuture.completedFuture(ImmutableMap.of("ID", "123")));
+        when(pdfGenerator.generate(any(), any())).thenReturn(CompletableFuture.supplyAsync(() -> new Byte[0]));
+
+        given(offenderApi.logon(any())).willReturn(CompletableFuture.completedFuture(JwtHelperTest.generateToken()));
+        given(offenderApi.getOffenderByCrn(any(), any())).willReturn(CompletableFuture.completedFuture(anOffenderWithNoContactDetails()));
+        given(offenderApi.getCourtAppearancesByCrn(any(), any()))
+                .willReturn(CompletableFuture.completedFuture(someCourtAppearances()));
+
     }
 
     @Test
@@ -60,7 +53,7 @@ public class SaveAsDraftWebTest extends WithIE8Browser {
 
         whenSaveAsDraftIsClicked();
 
-        verify(alfrescoDocumentStore, atLeastOnce()).updateExistingPdf(any(), any(), any(), any(), any());
+        verify(documentStore, atLeastOnce()).updateExistingPdf(any(), any(), any(), any(), any());
     }
 
     @Test
@@ -97,26 +90,5 @@ public class SaveAsDraftWebTest extends WithIE8Browser {
         offenderAssessmentPage.saveAsDraft();
     }
 
-    @Override
-    protected Application provideApplication() {
-        PdfGenerator pdfGenerator = mock(PdfGenerator.class);
-        when(pdfGenerator.generate(any(), any())).thenReturn(CompletableFuture.supplyAsync(() -> new Byte[0]));
-
-        OffenderApi offenderApi = mock(OffenderApi.class);
-        given(offenderApi.logon(any())).willReturn(CompletableFuture.completedFuture(JwtHelperTest.generateToken()));
-        given(offenderApi.getOffenderByCrn(any(), any())).willReturn(CompletableFuture.completedFuture(anOffenderWithNoContactDetails()));
-        given(offenderApi.getCourtAppearancesByCrn(any(), any()))
-            .willReturn(CompletableFuture.completedFuture(someCourtAppearances()));
-
-
-        return new GuiceApplicationBuilder().
-            overrides(
-                bind(PdfGenerator.class).toInstance(pdfGenerator),
-                bind(DocumentStore.class).toInstance(alfrescoDocumentStore),
-                bind(OffenderApi.class).toInstance(offenderApi),
-                bind(AnalyticsStore.class).toInstance(mock(AnalyticsStore.class)))
-            .configure("params.user.token.valid.duration", "100000d")
-            .build();
-    }
 
 }
